BSIS-937: Renamed ComponentBatchViewModels
Renamed ComponentBatchBasicViewModel to ComponentBatchViewModel and
ComponentBatchViewModel to ComponentBatchFullViewModel. Also changed
method names in the ViewModelFactory.
